==> 01_Research/03_feature_selection.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.linear_model import Lasso
from sklearn.feature_selection import SelectFromModel
pd.pandas.set_option("display.max_columns", None)

X_train = pd.read_csv("xtrain.csv")
X_test = pd.read_csv("xtest.csv")

X_train.head()

y_train = pd.read_csv("ytrain.csv")
y_test = pd.read_csv("ytest.csv")

y_train.head()

sel_ = SelectFromModel(Lasso(alpha=0.001, random_state=0))
sel_.fit(X_train, y_train)

sel_.get_support().sum()

sel_.get_support()

selected_feats = X_train.columns[(sel_.get_support())]

print(f"total features: {X_train.shape[1]}")
print(f"selected features: {len(selected_feats)}")
print(f"features with coefficients shrank to zero: {np.sum(sel_.estimator_.coef_ == 0)}")

pd.Series(selected_feats).to_csv("selected_features.csv", index=False)
