==> 01_Research/07_ferature_engineering_pipeline.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import joblib
from sklearn.linear_model import Lasso
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import MinMaxScaler, Binarizer
from sklearn.feature_selection import SelectFromModel
from feature_engine.imputation import (
    AddMissingIndicator,
    MeanMedianImputer,
    CategoricalImputer,
)

from feature_engine.encoding import (
    RareLabelEncoder,
    OrdinalEncoder,
)

from feature_engine.transformation import (
    LogTransformer,
    YeoJohnsonTransformer,
)

from feature_engine.selection import DropFeatures
from feature_engine.wrappers import SklearnTransformerWrapper

import preprocessors as pp
pd.pandas.set_option('display.max_columns', None)

data = pd.read_csv('train.csv')

# rows and columns of the data
print(data.shape)

# visualise the dataset
data.head()

data['MSSubClass'] = data['MSSubClass'].astype('O')

X_train, X_test, y_train, y_test = train_test_split(
    data.drop(['Id', 'SalePrice'], axis=1), # predictive variables
    data['SalePrice'], # target
    test_size=0.1, # portion of dataset to allocate to test set
    random_state=0, # we are setting the seed here
)

X_train.shape, X_test.shape

y_train = np.log(y_train)
y_test = np.log(y_test)

# categorical variables with NA in train set
CATEGORICAL_VARS_WITH_NA_FREQUENT = ['MasVnrType',
                                     'BsmtQual',
                                     'BsmtCond',
                                     'BsmtExposure',
                                     'BsmtFinType1',
                                     'BsmtFinType2',
                                     'Electrical',
                                     'GarageType',
                                     'GarageFinish',
                                     'GarageQual',
                                     'GarageCond']


CATEGORICAL_VARS_WITH_NA_MISSING = [
    'Alley', 'FireplaceQu', 'PoolQC', 'Fence', 'MiscFeature']


# numerical variables with NA in train set
NUMERICAL_VARS_WITH_NA = ['LotFrontage', 'MasVnrArea', 'GarageYrBlt']


TEMPORAL_VARS = ['YearBuilt', 'YearRemodAdd', 'GarageYrBlt']
REF_VAR = "YrSold"


# variables to log transform
NUMERICALS_LOG_VARS = ["LotFrontage", "1stFlrSF", "GrLivArea"]

NUMERICALS_YEO_VARS = ['LotArea']


BINARIZE_VARS = [
    'BsmtFinSF2', 'LowQualFinSF', 'EnclosedPorch',
    '3SsnPorch', 'ScreenPorch', 'MiscVal'
]

# variables to map
QUAL_VARS = ['ExterQual', 'ExterCond', 'BsmtQual', 'BsmtCond',
             'HeatingQC', 'KitchenQual', 'FireplaceQu',
             'GarageQual', 'GarageCond',
             ]

EXPOSURE_VARS = ['BsmtExposure']

FINISH_VARS = ['BsmtFinType1', 'BsmtFinType2']

GARAGE_VARS = ['GarageFinish']

FENCE_VARS = ['Fence']

# categorical variables to encode
CATEGORICAL_VARS = [
    'MSZoning',
    'Street',
    'Alley',
    'LotShape',
    'LandContour',
    'Utilities',
    'LotConfig',
    'LandSlope',
    'Neighborhood',
    'Condition1',
    'Condition2',
    'BldgType',
    'HouseStyle',
    'RoofStyle',
    'RoofMatl',
    'Exterior1st',
    'Exterior2nd',
    'MasVnrType',
    'Foundation',
    'Heating',
    'CentralAir',
    'Electrical',
    'Functional',
    'GarageType',
    'PavedDrive',
    'PoolQC',
    'MiscFeature',
    'SaleType',
    'SaleCondition',
    'MSSubClass']


QUAL_MAPPINGS = {'Po': 1, 'Fa': 2, 'TA': 3,
                 'Gd': 4, 'Ex': 5, 'Missing': 0, 'NA': 0}

EXPOSURE_MAPPINGS = {'No': 1, 'Mn': 2, 'Av': 3, 'Gd': 4}

FINISH_MAPPINGS = {'Missing': 0, 'NA': 0, 'Unf': 1,
                   'LwQ': 2, 'Rec': 3, 'BLQ': 4, 'ALQ': 5, 'GLQ': 6}

GARAGE_MAPPINGS = {'Missing': 0, 'NA': 0, 'Unf': 1, 'RFn': 2, 'Fin': 3}

FENCE_MAPPINGS = {'Missing': 0, 'NA': 0,
                  'MnWw': 1, 'GdWo': 2, 'MnPrv': 3, 'GdPrv': 4}

X_train[CATEGORICAL_VARS] = X_train[CATEGORICAL_VARS].astype('category')
X_test[CATEGORICAL_VARS] = X_test[CATEGORICAL_VARS].astype("category")

price_pipe = Pipeline([
    ("missing_imputation", CategoricalImputer(imputation_method="missing", variables=CATEGORICAL_VARS_WITH_NA_MISSING)),
    ("frequent_imputation", CategoricalImputer(imputation_method="frequent", variables=CATEGORICAL_VARS_WITH_NA_FREQUENT)),
    ("missing_indicator", AddMissingIndicator(variables=NUMERICAL_VARS_WITH_NA)),
    ("mean_imputation", MeanMedianImputer(imputation_method="mean", variables=NUMERICAL_VARS_WITH_NA)),
    ("elapsed_time", pp.TemporalVariableTransformer(variables=TEMPORAL_VARS, reference_variable=REF_VAR)),
    ("drop_features", DropFeatures(features_to_drop=[REF_VAR])),
    ("log", LogTransformer(variables=NUMERICALS_LOG_VARS)),
    ("yeojohnson", YeoJohnsonTransformer(variables=NUMERICALS_YEO_VARS)),
    ("binarizer", SklearnTransformerWrapper(transformer=Binarizer(threshold=0), variables=BINARIZE_VARS)),
    ("mapper_qual", pp.Mapper(variables=QUAL_VARS, mappings=QUAL_MAPPINGS)),
    ("mapper_exposure", pp.Mapper(variables=EXPOSURE_VARS, mappings=EXPOSURE_MAPPINGS)),
    ("mapper_finish", pp.Mapper(variables=FINISH_VARS, mappings=FINISH_MAPPINGS)),
    ("mapper_garage", pp.Mapper(variables=GARAGE_VARS, mappings=GARAGE_MAPPINGS)),
    ("mapper_fence", pp.Mapper(variables=FENCE_VARS, mappings=FENCE_MAPPINGS)),
    ("rare_label_encoder", RareLabelEncoder(tol=0.01, n_categories=1, variables=CATEGORICAL_VARS)),
    ("categorical_encoder", OrdinalEncoder(encoding_method="ordered", variables=CATEGORICAL_VARS))
])

# Check the data types of variables in CATEGORICAL_VARS
non_categorical_vars = [var for var in CATEGORICAL_VARS if X_train[var].dtype not in ['object', 'category']]

# Output the non-categorical variables
print("Non-categorical variables:", non_categorical_vars)


price_pipe.fit(X_train, y_train)

X_train = price_pipe.transform(X_train)
X_test = price_pipe.transform(X_test)

[var for var in X_train.columns if X_train[var].isnull().sum() > 0]

[var for var in X_test.columns if X_test[var].isnull().sum() > 0]

price_pipe.named_steps["frequent_imputation"].imputer_dict_

X_train.head()
